Petroleum revenue allocation  for 2020 - 2021
 Petroleum revenue allocation in the 2020 - 2021 financial year

Description: This dataset shows the revenue allocated to the petroleum sector in the 2020-2021 financial year. The data further lists the different areas for which the revenue was allocated.

Most recent changes: The dataset was last updated in December 2020

Languages: EN

Source: This data was obtained from the Ministry of Finance budget website by the African Centre for Media Excellence (ACME). For more information visit http://budget.go.ug/dataportal .